pyproject.toml
[tool.poetry]
name = "multion-reddit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
multion = "^1.3.6"
praw = "^7.7.1"
typer = "^0.12.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = 'multion_reddit.app:main'